Add files via upload
These are the header files to use, dependent on your operating system. It's a small, lightweight library that makes some of the lower level socket stuff easier. Not designed to be an entire networking framework, but something simple for those who want to design their net programs from the bottom up.